libの整理
