+++ src/lib/markdown/detector.ts
