+++ src/lib/markdown/spoiler.ts
@@ -1,0 +1,4 @@
+export const toggleSpoiler = (element: HTMLElement) => {
+  const $spoiler = element.closest('.spoiler')
+  $spoiler?.toggleAttribute('shown')
+}
