+++ src/store/domain/messagesView/actions.ts
@@ -3,7 +3,7 @@ import { moduleActionContext } from '@/store'
 import { messagesView } from './index'
 import { ChannelId, MessageId, ClipFolderId } from '@/types/entity-ids'
 import { ChannelViewer, ChannelViewState, Message } from '@traptitech/traq'
-import { render } from '@/lib/markdown'
+import { render } from '@/lib/markdown/markdown'
 import apis from '@/lib/apis'
 import { changeViewState } from '@/lib/websocket'
 import { ActionContext } from 'vuex'
