+++ tests/unit/lib/markdown/internalLinkEmbedder.spec.ts
@@ -1,4 +1,8 @@
-import { ReplaceGetters, replace, Entity } from '@/lib/internalLinkEmbedder'
+import {
+  ReplaceGetters,
+  replace,
+  Entity
+} from '@/lib/markdown/internalLinkEmbedder'
 
 const users = {
   'dfdff0c9-5de0-46ee-9721-2525e8bb3d45': {
