+++ tests/unit/lib/markdown/spoiler.spec.ts
@@ -1,4 +1,4 @@
-import { toggleSpoiler } from '@/lib/markdown'
+import { toggleSpoiler } from '@/lib/markdown/spoiler'
 
 describe('toggleSpoiler', () => {
   it('can toggle to shown', () => {
